requirements.txt
torch
torchvision
torchaudio
plotly
pandas
scikit-learn
histolab
openslide-python
py-params
matplotlib
torchmetrics
wandb
